requirements.txt
python-dotenv
git+https://github.com/Rapptz/discord.py